fix bin and tests
